Documentation for entmootopti and tests added, gurobi_env moved to utils
